+++ echoshader/plots/map.py
@@ -6,7 +6,7 @@ import pandas
 import xarray
 from pyproj import Transformer
 
-from .utils import EPSG_coordsys, EPSG_mercator, gram_opts
+from ..utils import EPSG_coordsys, EPSG_mercator, gram_opts
 
 opt_tile = geoviews.opts(tools=["box_select"])
